% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_imputation.R
\name{evaluate_imputation}
\alias{evaluate_imputation}
\title{Imputation Evaluation}
\usage{
evaluate_imputation(original_matrix, imputed_matrix)
}
\arguments{
\item{original_matrix}{Original matrix}

\item{imputed_matrix}{Matrix for which missigness has been for example artificially created and then imputed}
}
\value{
matrix containing rmse, mae, rmse_cv and mape for each column
}
\description{
This function calculates for evaluation metrics between matrix with no missigness and matrix with imputed values. Four metrics will be returned: root mean square error, mean absolute error, coeficient of variation of root mean square error and mean absolute percentage error for each column in the matrices.
}
\examples{
mat = as.matrix(iris[,1:4])
mis_mat = generate_na(mat , 0.3)
imp_mat = impute_xgboost(mis_mat)
evaluate_imputation(mat, imp_mat)

}
