% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_na.R
\name{generate_na}
\alias{generate_na}
\title{Adds Missing Values to a Data Set}
\usage{
generate_na(data, p = 0.1, seed = NULL)
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{p}{Proportion of missing values to approximately add to each column of \code{data}.}

\item{seed}{An integer seed.}
}
\value{
\code{data} with missing values.
}
\description{
Takes a data frame and replaces randomly part of the values by missing values.
}
\examples{
head(generateNA(iris))
}
