% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_decode.R
\name{one_hot_decode}
\alias{one_hot_decode}
\title{Decode one hot encoding}
\usage{
one_hot_decode(mat, verbose = 1)
}
\arguments{
\item{mat}{A matrix created by one_hot_encode-function and passed through impute_xgboost}

\item{verbose}{Should decoding progress be printed}
}
\value{
A data.frame or data.table depending on the original class given to one_hot_encode-function. Column classes will be retained.
}
\description{
Decodes one hot encoding generated by one_hot_encode-function utilizing matrix attributes generated in one_hot_encode-function.
}
\examples{

mat = one_hot_encode(iris)
mis_mat = generate_na(mat)
imp_mat = impute_xgboost(mis_mat)
imp_iris = one_hot_decode(imp_mat)

}
