% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_univariate.R
\name{impute_univariate}
\alias{impute_univariate}
\title{Univariate Imputation}
\usage{
impute_univariate(x)
}
\arguments{
\item{x}{A vector of any type possibly containing missing values.}
}
\value{
A vector of the same length and type as \code{x} but without missing values.
}
\description{
Fills missing values of a vector of any type by sampling
with replacement from the non-missing values. Requires at least one
non-missing value to run.
}
\examples{
impute_univariate(c(NA, 0, 1, 0, 1))
impute_univariate(c("A", "A", NA))
impute_univariate(as.factor(c("A", "A", NA)))
}
